Move Options into <PolicyType> namespaces
