+++ src/Polly.Contrib.Simmy.Specs/Behavior/InjectBehaviourAsyncWithOptionsSpecs.cs
@@ -2,6 +2,7 @@
 using Polly.Contrib.Simmy.Utilities;
 using System;
 using System.Threading.Tasks;
+using Polly.Contrib.Simmy.Behavior;
 using Xunit;
 
 namespace Polly.Contrib.Simmy.Specs.Behavior
