+++ src/Polly.Contrib.Simmy.Specs/Behavior/InjectBehaviourWithOptionsSpecs.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using FluentAssertions;
+using Polly.Contrib.Simmy.Behavior;
 using Polly.Contrib.Simmy.Utilities;
 using Xunit;
