+++ src/Polly.Contrib.Simmy.Specs/Latency/InjectLatencyAsyncWithOptionsSpecs.cs
@@ -3,6 +3,7 @@ using System.Diagnostics;
 using System.Threading;
 using System.Threading.Tasks;
 using FluentAssertions;
+using Polly.Contrib.Simmy.Latency;
 using Polly.Contrib.Simmy.Utilities;
 using Polly.Utilities;
 using Xunit;
