+++ src/Polly.Contrib.Simmy.Specs/Latency/InjectLatencyTResultAsyncWithOptionsSpecs.cs
@@ -7,6 +7,7 @@ using Polly.Contrib.Simmy.Utilities;
 using Polly.Utilities;
 using Xunit;
 using FluentAssertions;
+using Polly.Contrib.Simmy.Latency;
 using Polly.Timeout;
 
 namespace Polly.Contrib.Simmy.Specs.Latency
