+++ src/Polly.Contrib.Simmy/Behavior/InjectBehaviourOptions.cs
@@ -1,8 +1,7 @@
 ﻿using System;
 using System.Threading;
-using Polly.Contrib.Simmy.Behavior;
 
-namespace Polly.Contrib.Simmy
+namespace Polly.Contrib.Simmy.Behavior
 {
     /// <summary>
     /// Options used to configure an <see cref="AsyncInjectBehaviourPolicy"/>
