+++ src/Polly.Contrib.Simmy/Latency/InjectLatencyAsyncOptions.cs
@@ -1,9 +1,8 @@
 ﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
-using Polly.Contrib.Simmy.Latency;
 
-namespace Polly.Contrib.Simmy
+namespace Polly.Contrib.Simmy.Latency
 {
     /// <summary>
     /// Options used to configure an <see cref="AsyncInjectLatencyPolicy"/>
