+++ src/Polly.Contrib.Simmy/Outcomes/InjectFaultOptions.cs
@@ -1,8 +1,7 @@
 ﻿using System;
 using System.Threading;
-using Polly.Contrib.Simmy.Outcomes;
 
-namespace Polly.Contrib.Simmy
+namespace Polly.Contrib.Simmy.Outcomes
 {
     /// <summary>
     /// Options used to configure an <see cref="InjectOutcomePolicy"/>
